+++ instagram_app/serializers/image.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from instagram_app.models import Images
+
+class ImagesSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Images
+        fields = ['image']
