+++ instagram_app/urls.py
@@ -1,7 +1,9 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from instagram_app.views import (
-    PostsView, UserView, LoginView, LogoutView, LikeView, CommentView, ProfileStoriesView)
+from instagram_app.views.user import UserView, ProfileStoriesView, LoginView, LogoutView
+from instagram_app.views.post import PostsView
+from instagram_app.views.comment import CommentView
+from instagram_app.views.like import LikeView
 
 
 urlpatterns = [
